[ETFE-3134] - PR comments
